ias15 using size_t

+++ rebound/integrators/ias15.py
@@ -87,6 +87,6 @@ class IntegratorIAS15(ctypes.Structure):
                 ("_e", reb_dp7),
                 ("_br", reb_dp7),
                 ("_er", reb_dp7),
-                ("_map", ctypes.POINTER(ctypes.c_int)),
-                ("_map_allocated_n", ctypes.c_uint),
+                ("_map", ctypes.POINTER(ctypes.c_size_t)),
+                ("_map_allocated_n", ctypes.c_size_t),
                 ]
